Replace CreditCardFields with CreditCardLoading
CreditCardFields was only being used for a disabled loader at this point.
